% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nz_add_metadata.R
\name{nz_add_metadata}
\alias{nz_add_metadata}
\title{Add metadata}
\usage{
nz_add_metadata(DSN, table_name, description = NA, legend = NA)
}
\arguments{
\item{DSN}{DSN object extracted from nz_init}

\item{table_name}{name of the table}

\item{description}{table description. if you put it NA it will not update description and legend.}

\item{legend}{legend of the layer.}
}
\value{

}
\description{
Add metadata
}
\section{Functions}{
\itemize{
\item \code{nz_add_metadata}: metadata rows to the database. it will remove all the exisiting records.
To be able to render a layer you have to call this function on the table. Everytime you update values for
a key in the table you also need to call this function
}}

\examples{
\dontrun{
DSN <- nz_init("NZSQLF","SPATIAL_SCHEMA")
nz_add_metadata(DSN,"NLULCDATA")
}
}
