% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeClass.R
\name{convert_datetime_to_tid}
\alias{convert_datetime_to_tid}
\title{convert_datetime_to_tid
Convert date and time to TID scale.}
\usage{
convert_datetime_to_tid(datetime, scale)
}
\arguments{
\item{datetime}{a date and time value in string format like '2016-01-02 00:00:00'. It must be
between "1000-01-01 00:00:00" and "3000-01-01 00:00:00"}

\item{scale}{the scale for TID. Must be one the following items
'1000y','500y','100y','50y','10y','1y','m','w','d','h','min','s'}
}
\value{
TID in an Integer format
}
\description{
convert_datetime_to_tid
Convert date and time to TID scale.
}
\examples{
convert_datetime_to_tid('1y','1980-01-01 00:00:00')
}
