% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataImporterClass.R
\name{import_to_db}
\alias{import_to_db}
\title{import_to_db}
\usage{
import_to_db(
  DSN,
  directory,
  table_name,
  value_type = "varchar",
  createTable = T
)
}
\arguments{
\item{DSN}{The NZ DSN}

\item{directory}{Directory of CSV files}

\item{table_name}{Name of the table to import data into}

\item{value_type}{The type of Value possible options
float, varchar, integer, bigint}

\item{createTable}{Either make a new table and drop table if exists or make table}
}
\value{

}
\description{
import_to_db
}
\examples{
import_to_db()
}
\keyword{import}
\keyword{netezza,}
