% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nz_convert_df_to_dggs.R
\name{nz_convert_df_to_dggs}
\alias{nz_convert_df_to_dggs}
\title{dataframe to dggs}
\usage{
nz_convert_df_to_dggs(DGGID, TID, df, SaveIn, convertKeys = "all")
}
\arguments{
\item{DGGID}{A dataframe column containing DGGIDs}

\item{TID}{An integer or dataframe column}

\item{df}{dataframe with the remaining keys. Remove DGGID/TID columns from it}

\item{SaveIn}{The directory path to store data there. It must exist}

\item{convertKeys}{Keys that are supposed to be converted. \strong{all} means converting all the
keys. \strong{NA} means does not convert any keys.Use a vector of column names to only convert specific keys}
}
\value{

}
\description{
Converts a dataframe with a DGGID/TID column and set of other columns (keys) to csv
datasets to be able to import them into netezza. You can import csv files to the database using nz_import_dir_to_db function
}
\examples{
\dontrun{
csv<-read.csv2("D:\\\\UserData\\\\Majid\\\\Downloads\\\\mpb-ev.csv",sep=",")
head(csv)
nz_convert_df_to_dggs(csv$DGGID,-1,csv,"E:\\\\home\\\\crobertson")
#importing to the db
DSN <- nz_init("NZSQLF","SPATIAL_SCHEMA")
nz_import_dir_to_db(DSN,"/home/crobertson/import/","mbpev","varchar(100)")

}
}
