% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nz_convert_points_to_dggs.R
\name{nz_convert_points_key_value_to_dggs}
\alias{nz_convert_points_key_value_to_dggs}
\title{Convert lat,lon, tid, key,value to a dggs data model}
\usage{
nz_convert_points_key_value_to_dggs(lat, lon, tid, res, key, value, save_in)
}
\arguments{
\item{lat}{y coordinates of the points (EPSG:4326)}

\item{lon}{x coordinates of the points (EPSG:4326)}

\item{tid}{TID value, A single integer for all data or a list of integer values}

\item{res}{Resolution of the final dggs cells, a single integer value or a list of integer values}

\item{key}{a single string as key}

\item{value}{a list of value with the length equal to the lat}

\item{save_in}{the directory to store outpur csv files.make sure that this directory exist, you have write permission and engough disc space}

\item{df}{A dataframe of keys. The number of rows of this dataframe must be equal to the lat,lon rows}
}
\value{

}
\description{
Convert lat,lon, tid, key,value to a dggs data model
}
\examples{
\dontrun {
r <- read.csv('D:/Bathurst_caribou_collars.csv')
nz_convert_points_key_value_to_dggs(r$Latitude,r$Longitude,100,10,"key",value,"C:/result")
}
}
