% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nz_convert_raster_to_dggs_by_centroid.R
\name{nz_convert_raster_to_dggs_by_centroid}
\alias{nz_convert_raster_to_dggs_by_centroid}
\title{Raster to dggs}
\usage{
nz_convert_raster_to_dggs_by_centroid(rast, centroids, key, tid)
}
\arguments{
\item{rast}{Raster Object exported from Raster package}

\item{centroids}{centroids dg made by nz_make_centroids_df_from_csv function}

\item{key}{key}

\item{tid}{tid value}
}
\value{
A dataframe object
}
\description{
Convert Raster object to dggs by centroids. You have to save csv file by yourself.Like using which write.csv(df2,name,sep=';', row.names=FALSE,)
}
