% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nz_dplyr_to_netezza.R
\name{nz_dplyr_to_netezza}
\alias{nz_dplyr_to_netezza}
\title{dplyr to Netezza table}
\usage{
nz_dplyr_to_netezza(DSN, dplyr, outputTable, dropIfExist = F)
}
\arguments{
\item{DSN}{A DSN object exported from nz_init function}

\item{dplyr}{a dplyr object}

\item{outputTable}{name of the output table to store data}

\item{dropIfExist}{Drop if the outputTable does exist. Default is False. Use it with cautious since it removes
data in the existing outputTable permanently}
}
\value{
A dplyr object from outputTable
}
\description{
Save a dplyr object as netezza table. It can drop table when it already exists.
}
\examples{
\dontrun{
DSN <- nz_init("NZSQL_F","ADMIN")
mbp <- nz_table_as_dplyr(DSN,"MPB")
head(mbp)
mbp2 <- nz_dplyr_to_netezza(DSN,mbp,"mbp2")
head(mbp2)
}
}
