% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{nz_export_centroids_from_db}
\alias{nz_export_centroids_from_db}
\title{Export centroids from Netezza}
\usage{
nz_export_centroids_from_db(DSN_NAME, output_directory, resolution)
}
\arguments{
\item{DSN_NAME}{NZODBC DSN name}

\item{output_directory}{The output directory to store csv file (like F:\\data\\store)it should be available on the disc}

\item{resolution}{the resolution of the data}
}
\value{

}
\description{
Export centroids from Netezza
}
