% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataImporterClass.R
\name{nz_find_duplicates}
\alias{nz_find_duplicates}
\title{Title}
\usage{
nz_find_duplicates(DSN_NAME, table)
}
\arguments{
\item{DSN_NAME}{Nz DSN name}

\item{table}{table name in format of   SPATIAL_SCHEMA.TESTMEE}
}
\value{
a sqlQuery object
}
\description{
Title
}
