% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataImporterClass.R
\name{nz_import_dir_to_db}
\alias{nz_import_dir_to_db}
\title{import directory}
\usage{
nz_import_dir_to_db(
  DSN,
  directory,
  table_name,
  value_type = "varchar",
  createTable = T
)
}
\arguments{
\item{DSN}{The NZ DSN object made by nz_init}

\item{directory}{Directory of CSV files}

\item{table_name}{Name of the table to import data into}

\item{value_type}{The type of Value possible options
float, varchar, integer, bigint}

\item{createTable}{Either make a new table and drop table if exists or append data to the existing table}
}
\value{

}
\description{
Import a directory into database. The directory must have a set of csv files
}
\examples{
\dontrun{nz_import_dir_to_db()}
}
\keyword{import}
\keyword{netezza,}
