% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataImporterClass.R
\name{nz_import_file_to_db}
\alias{nz_import_file_to_db}
\title{Import a file to db}
\usage{
nz_import_file_to_db(
  DSN,
  file_path,
  table_name,
  value_type = "varchar(100)",
  createTable = T,
  max_errors = 2
)
}
\arguments{
\item{DSN}{object extracted from nz_init function}

\item{file_path}{the csv file path with following format "VALUE","DGGID","TID","KEY"}

\item{value_type}{The type of Value possible options
float, varchar, integer, bigint}

\item{createTable}{Either make a new table and drop table if exists or append data to the existing table}

\item{max_errors}{The maximumn number of rows in the csv which can include error in their values}
}
\value{

}
\description{
Import a single csv file to the netezza database. CSV columns must be as with following format "VALUE","DGGID","TID","KEY"
}
