% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nz_init.R
\name{nz_init}
\alias{nz_init}
\title{make nz db object}
\usage{
nz_init(DSN_NAME, SCHEMA)
}
\arguments{
\item{DSN_NAME}{The DSN name for connection}

\item{SCHEMA}{The database Schema}
}
\value{
an nz db object
}
\description{
Makes an Netezza db object for the user. Most of functions that need to interact with th
netezza database need this object as input.
}
\examples{
obj <- nz_init("NZSQL","SPATIAL_SCHEMA")
}
