% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nz_keylookup_getkeys.R
\name{nz_keylookup_getkeys}
\alias{nz_keylookup_getkeys}
\title{keylookup table keys}
\usage{
nz_keylookup_getkeys(DSN, table = NA)
}
\arguments{
\item{DSN}{Nz information object}

\item{table}{The teble to get keys from. It can also be NA so it will return entire keylookup table. It only queries keys which ends with __ED}
}
\value{
a sqlQuery object
}
\description{
get the keys in the keylookup table
}
\examples{
\dontrun{
nz_keylookup_getkeys(DSN,table = "TEST_MBPEV")
}
}
