% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nz_table_as_dplyr.R
\name{nz_table_as_dplyr}
\alias{nz_table_as_dplyr}
\title{Netezza Table To dplyr}
\usage{
nz_table_as_dplyr(DSN, tableName)
}
\arguments{
\item{DSN}{DSN object extracted from nz_init function}

\item{tableName}{the name of the table to read}
}
\value{
A dplyr object
}
\description{
Converts a Netezza table to a dplyr object.
}
\examples{
\dontrun{
DSN <- nz_init("NZSQL_F","ADMIN")
mbp <- nz_table_as_dplyr(DSN,"MPB")
head(mbp)
}

}
