% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeClass.R
\name{TimeClass}
\alias{TimeClass}
\title{setBoundLimit
Define TID bounds}
\description{
setBoundLimit
Define TID bounds

setBoundLimit
Define TID bounds
}
\examples{

## ------------------------------------------------
## Method `TimeClass$setBoundLimit`
## ------------------------------------------------

dateTimeToPOT
Convert date and time to TID scale.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-setBoundLimit}{\code{TimeClass$setBoundLimit()}}
\item \href{#method-dateTimeToPOT}{\code{TimeClass$dateTimeToPOT()}}
\item \href{#method-getP}{\code{TimeClass$getP()}}
\item \href{#method-getLeapYears}{\code{TimeClass$getLeapYears()}}
\item \href{#method-POTToDateTime}{\code{TimeClass$POTToDateTime()}}
\item \href{#method-clone}{\code{TimeClass$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setBoundLimit"></a>}}
\if{latex}{\out{\hypertarget{method-setBoundLimit}{}}}
\subsection{Method \code{setBoundLimit()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeClass$setBoundLimit(start, end)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{start}}{the begining of TID strptime( '01-01-1000 00:00:00', "\%d-\%m-\%Y \%H:\%M:\%S")}

\item{\code{end}}{the end of TID strptime( '01-01-3000 00:00:00', "\%d-\%m-\%Y \%H:\%M:\%S")}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{

}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{dateTimeToPOT
Convert date and time to TID scale.
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dateTimeToPOT"></a>}}
\if{latex}{\out{\hypertarget{method-dateTimeToPOT}{}}}
\subsection{Method \code{dateTimeToPOT()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeClass$dateTimeToPOT(date, scale)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{scale}}{the scale for TID. Must be one the following items
'1000y','500y','100y','50y','10y','1y','m','w','d','h','min','s'}

\item{\code{datetime}}{a date and time value in string format like '2016-01-02 00:00:00'. It must be
between "1000-01-01 00:00:00" and "3000-01-01 00:00:00"}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
TID in an Integer format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getP"></a>}}
\if{latex}{\out{\hypertarget{method-getP}{}}}
\subsection{Method \code{getP()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeClass$getP()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getLeapYears"></a>}}
\if{latex}{\out{\hypertarget{method-getLeapYears}{}}}
\subsection{Method \code{getLeapYears()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeClass$getLeapYears()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-POTToDateTime"></a>}}
\if{latex}{\out{\hypertarget{method-POTToDateTime}{}}}
\subsection{Method \code{POTToDateTime()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeClass$POTToDateTime(value)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TimeClass$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
