% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_to_db}
\alias{import_to_db}
\title{import_to_db}
\usage{
import_to_db(directory, table_name, append = T, drop = F)
}
\arguments{
\item{directory}{Directory of CSV files}

\item{table_name}{Name of the table to import data into}

\item{append}{add data to an existing table, if it is False we will make a new table}

\item{drop}{drops table while creating a new table, Default is False}
}
\value{

}
\description{
import_to_db
}
\examples{
import_to_db()
}
\keyword{import}
\keyword{netezza,}
