% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nz_export_centroids_from_db.R
\name{nz_export_centroids_from_db}
\alias{nz_export_centroids_from_db}
\title{Export centroids}
\usage{
nz_export_centroids_from_db(DSN, output_directory, resolution)
}
\arguments{
\item{DSN}{NZODBC DSN object made by nz_init function}

\item{output_directory}{The output directory to store csv file (like F:\\data\\store)it should be available on the disc}

\item{resolution}{the resolution of the data}
}
\value{

}
\description{
Export centroids from Netezza FinalGrid table for specific resolution.
}
