% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nz_find_duplicates.R
\name{nz_find_duplicates}
\alias{nz_find_duplicates}
\title{Find duplicates}
\usage{
nz_find_duplicates(DSN, table, limit = T)
}
\arguments{
\item{DSN}{Nz information object}

\item{table}{table name in format of}

\item{limit}{Whether to show all the duplicates or only show 100 result. It can cause performance issue on large
tables if you set it to FALSE. Default is TRUE}
}
\value{
a sqlQuery object
}
\description{
Find duplicate cells in a table. If the returned dataframe is empty it does not have any duplicate cells.
}
