% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataImporterClass.R
\name{nz_import_file_to_db}
\alias{nz_import_file_to_db}
\title{Title}
\usage{
nz_import_file_to_db(
  DSN,
  file_path,
  table_name,
  value_type = "varchar",
  createTable = T
)
}
\arguments{
\item{file_path}{the csv file path with following format "VALUE","DGGID","TID","KEY"}

\item{value_type}{The type of Value possible options
float, varchar, integer, bigint}

\item{createTable}{Either make a new table and drop table if exists or make table}
}
\value{

}
\description{
Title
}
